vars: extend signature of a checkout function
Add version of generic_checkout() function accepting map instead
of list of arguments: it can be more convenient and human readable
to call it with named arguments when list of arguments is long and
some of them may be omitted.
Add argument allowing to disable polling SCM - it is useful when
the same repository is cloned multiple times.

OL-Redmine-Id: 12984
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>