vars: improve handling of jobs providing revisions to build
Rename 'update_job' parameter into 'get_revs_from'. Allow it to take
a list of jobs instead of a single job name - it is possible that
there are a few jobs, each checking different repository.
Implement a function allowing to construct triggers from a list of
jobs - it is useful when a job checking one repository depends on
a job checking another repository.

OL-Redmine-Id: 12984
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>